chore: move recalculate_workflow_tags to data layer
